+++ Tests/MassTransit.RabbitMQ/MassTransit.RabbitMQ.Application/IntegrationEvents/NamingOverrides/OverrideMessageStandardSubscribeEvent.cs
@@ -1,0 +1,17 @@
+using Intent.RoslynWeaver.Attributes;
+
+[assembly: DefaultIntentManaged(Mode.Fully)]
+[assembly: IntentTemplate("Intent.Eventing.Contracts.IntegrationEventMessage", Version = "1.0")]
+
+namespace MassTransit.RabbitMQ.Services
+{
+    public record OverrideMessageStandardSubscribeEvent
+    {
+        public OverrideMessageStandardSubscribeEvent()
+        {
+            Message = null!;
+        }
+
+        public string Message { get; init; }
+    }
+}