+++ Modules/Intent.Modules.AspNetCore.Logging.Serilog/release-notes.md
@@ -1,4 +1,4 @@
-### Versopm 5.1.3
+### Version 5.1.3
 
 - Improvement: Adding the Application Insights sink will only "force" you to specify a connection string when a environment is added to your appsettings.json that is non-Development. Gives a hassle-free out the box experience when running your application while ensuring that when deploying to a non-Development environment it ensures you specify a connection string.